Move crate to crate::json
